shop/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.3
django-cachalot==2.7.0
django-celery-beat==2.7.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-rest-registration==0.9.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
drf-yasg==1.21.8
flower==2.0.1
humanize==4.11.0
inflection==0.5.1
kombu==5.4.2
packaging==24.2
pillow==11.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
